%% getRI function %%
\name{getRI}
\alias{getRI}

\title{Generate a calibration function}

\description{This function generates a calibration function which can be used to calculate the retention index of a compound, given its retention time together with the retention indexes of RI standards and their retention times.}

\usage{
getRI(RItable = data.frame())
}

%- maybe also 'usage' for other objects documented here.
\arguments{
    \item{RItable}{a data frame of retention times and indexes of RI standardsPeaks, e.g. generated by calling \code{getRIStandard}.}
}

\details{The input of this function is an RItable. RItable is the table of RTs and correponding RIs of RI standards measured using an RI run. The user can get the RItable using the \code{getRIStandard} function. The output of this function is a function which can be used for retention index calculation providing a retention time.
}

\value{
A calibration function which accepts a retention time and outputs corresponding retention index.
}

\seealso{\code{\link{getRIStandard}}}

\references{\url{http://omics.georgetown.edu/SIMAT.html}}

\author{Mo Nezami Ranjbar}

%\note{  No notes for this function either }

\examples{
    # load retention index table from RI standards
    data(RItable)

    # create a calibration function
    calcRI <- getRI(RItable)
    
    # perform RI calibration for a certain RT = 7.32min
    calcRI(12.32)
}