\name{SIMAT-package}
\alias{SIMAT-package}
\alias{SIMAT}
\docType{package}
\title{
This package is for processing GC-SIM-MS data
}
\description{
This package provides a pipeline for analysis of GC-MS data acquired in selected ion monitoring (SIM) mode. The tool also provides a guidance in choosing appropriate fragments for the targets of interest by using an optimization algorithm. This is done by considering overlapping peaks from a provided library by the user.
}
\details{
\tabular{ll}{
Package: \tab SIMAT\cr
Type: \tab Package\cr
Version: \tab 0.99.2\cr
Date: \tab 2015-03-01\cr
License: \tab GPL-2\cr
}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Mo R. Nezami Ranjbar,
%~~ The author and/or maintainer of the package ~~
}

\references{
M. R. Nezami Ranjnar, C. Di Poto, Y. Wang, and H. W. Ressom, "SIMAT: GC-SIM-MS Data Analysis Tool", submitted to BMC Bioinformatics.
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
%~~ documentation directory ~~
\keyword{ package, gas chromatography, mass spectrometry, spectral library, targeted analysis }

\seealso{
\url{http://omics.georgetown.edu/SIMAT.html}
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}

%% 